expansion完成
